\name{read_metadata}

\alias{read_metadata}

\title{
Get the dataset level meta-data
}

\description{
Read the meta-data file associated with a data set (from the associated json file). 
}

\usage{
read_metadata(uri, path)
}
 
\arguments{
  \item{uri}{character. The URI of the dataset}
  \item{path}{the path to your local copy of the Carob repo}
} 

\value{
json
}

