\name{extract_metadata}

\alias{extract_metadata}

\title{
Get the dataset level meta-data
}

\description{
Extract meta-data associated with a data set
}

\usage{
extract_metadata(js, uri)
}
 
\arguments{
  \item{js}{JSON object created with \code{\link{read_metadata}}}
  \item{uri}{character. The URI of the dataset}
} 

\value{
data.frame
}

