\name{extract_metadata}

\alias{extract_metadata}

\title{
Extract dataset level meta-data
}

\description{
Extract some meta-data associated with a data set
}

\usage{
extract_metadata(uri, path)
}
 
\arguments{
  \item{uri}{character. The URI of the dataset}
  \item{path}{the path to your local copy of the Carob repo}
} 

\value{
data.frame
}

